increase ternary table max_size

+++ src/hardware_dep/shared/data_plane/ternary_naive.c
@@ -14,7 +14,7 @@
 #include "ternary_naive.h"
 
 ternary_table*
-naive_ternary_create(uint8_t keylen, uint8_t max_size)
+naive_ternary_create(uint8_t keylen, uint64_t max_size)
 {
     ternary_table* t = malloc(sizeof(ternary_table));
     t->entries = malloc(sizeof(ternary_entry)*max_size);
